refactor: :recycle: detach logic to useStaffDetails hook
